feat: add config node

+++ include/aruku/walking/node/walking_manager.hpp
@@ -73,7 +73,7 @@ private:
 
   std::vector<tachimawari::joint::Joint> joints;
   std::array<double, 18> inital_joints;
-  std::array<int, 18> joints_direction;
+  std::array<double, 18> joints_direction;
 
   keisan::Point2 position;
